apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API service"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.2",
    "dspy>=3.0.4",
    "fastapi>=0.128.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "openai>=1.59.0",
    "RestrictedPython>=6.0",
    "datasets>=3.2.0",
    "rank-bm25>=0.2.2",
    "pydantic-settings>=2.7.1",
    "websockets>=12.0",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.3",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
    "ty>=0.0.8",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []
